test2

+++ src/main/java/com/taskplanner/taskplanner/repositories/TaakRepository.java
@@ -17,5 +17,5 @@ public interface TaakRepository extends JpaRepository<Task, Long> {
     @Modifying
     @Transactional
     @Query("update Task t set t.name = ?1 , t.description = ?2, t.dueDate = ?3 where t.id = ?4")
-    void editTaak(String name, String description, LocalDateTime dueDate, int id);
+    void editTaak(String name, String description, LocalDateTime dueDate, long id);
 }
